[수정] UserService, GameService, PieceService를 GameService로 통합
